prov/netdir Remove excessive files - netdir_attr.c and netdir_err.h
