+++ prov/netdir/src/netdir_cntr.c
@@ -34,7 +34,6 @@
 
 #include "netdir.h"
 #include "netdir_ov.h"
-#include "netdir_err.h"
 #include "netdir_iface.h"
 
 #include "rdma/fabric.h"
