+++ prov/netdir/src/netdir_ep.c
@@ -37,7 +37,6 @@
 
 #include "netdir.h"
 #include "netdir_ov.h"
-#include "netdir_err.h"
 #include "netdir_log.h"
 #include "netdir_util.h"
 #include "netdir_iface.h"
