+++ prov/netdir/src/netdir_ep_msg.c
@@ -38,7 +38,6 @@
 #include "netdir.h"
 #include "netdir_ov.h"
 #include "netdir_cq.h"
-#include "netdir_err.h"
 #include "netdir_log.h"
 #include "netdir_iface.h"
 #include "netdir_unexp.h"
