+++ prov/netdir/src/netdir_eq.c
@@ -34,7 +34,6 @@
 
 #include "netdir.h"
 #include "netdir_ov.h"
-#include "netdir_err.h"
 #include "netdir_log.h"
 #include "netdir_iface.h"
