+++ prov/netdir/src/netdir_mr.c
@@ -35,7 +35,6 @@
 #include "netdir.h"
 
 #include "netdir_ov.h"
-#include "netdir_err.h"
 #include "netdir_iface.h"
 
 #include "ofi.h"
