+++ prov/netdir/src/netdir_unexp.c
@@ -40,7 +40,6 @@
 #include "netdir.h"
 #include "netdir_cq.h"
 #include "netdir_log.h"
-#include "netdir_err.h"
 #include "netdir_util.h"
 #include "netdir_queue.h"
 #include "netdir_iface.h"
